Update main.cpp
Fixed a bug that would make the program crash when deleting the pizza array after executing the medium.in file, and that would not divide the medium pizza the same each time. The pizza array was created as totalColumnsXtotalRows array, not as totalRowsXtotalColumns as it should have been. Which is why when we called pizza[i][j], pizza[i] was actually a column of the matrix, not a row.